+++ python/tk_vred/__init__.py
@@ -11,4 +11,4 @@
 from .menu_generation import VREDMenuGenerator
 from .dock_widget import DockWidget
 from .vred_py import VREDPy
-from .progress_bar_widget import VREDFileIOProgressWidget
+from .progress_widget import VREDFileIOProgressWidget
